Refactor changelog
